+++ packages/app-page-builder/src/editor/plugins/elementSettings/link/HrefSettings.tsx
@@ -5,7 +5,7 @@ import { Switch } from "@webiny/ui/Switch";
 import { Form } from "@webiny/form";
 import { validation } from "@webiny/validation";
 import { withActiveElement } from "~/editor/components";
-import { DelayedOnChange } from "~/editor/components/DelayedOnChange";
+import { DelayedOnChange } from "@webiny/ui/DelayedOnChange";
 import { useEventActionHandler } from "~/editor/hooks/useEventActionHandler";
 import { UpdateElementActionEvent } from "~/editor/recoil/actions";
 import { PbEditorPageElementSettingsRenderComponentProps, PbEditorElement } from "~/types";
