+++ packages/app-page-builder/src/pageEditor/config/blockEditing/SearchBlocks.tsx
@@ -36,7 +36,7 @@ import { useEventActionHandler } from "~/editor/hooks/useEventActionHandler";
 import { useKeyHandler } from "~/editor/hooks/useKeyHandler";
 import { UpdateElementActionEvent } from "~/editor/recoil/actions";
 import { createBlockElements } from "~/editor/helpers";
-import { DelayedOnChange } from "~/editor/components/DelayedOnChange";
+import { DelayedOnChange } from "@webiny/ui/DelayedOnChange";
 import { blocksBrowserStateAtom } from "~/pageEditor/config/blockEditing/state";
 
 const allBlockCategory: PbEditorBlockCategoryPlugin = {
